Feat: Drag the node to the group node
