+++ web/.umirc.ts
@@ -34,7 +34,7 @@ export default defineConfig({
   proxy: [
     {
       context: ['/api', '/v1'],
-      target: 'http://127.0.0.1:9380/',
+      target: 'http://127.0.0.1:9456/',
       changeOrigin: true,
       ws: true,
       logger: console,
