+++ web/src/pages/flow/canvas/node/iteration-node.tsx
@@ -1,12 +1,40 @@
 import { useTheme } from '@/components/theme-provider';
 import { cn } from '@/lib/utils';
 import { CirclePower } from 'lucide-react';
-import { Handle, NodeProps, Position } from 'reactflow';
+import { Handle, NodeProps, NodeResizeControl, Position } from 'reactflow';
 import { NodeData } from '../../interface';
 import useGraphStore from '../../store';
 import { LeftHandleStyle, RightHandleStyle } from './handle-icon';
 import styles from './index.less';
 import NodeHeader from './node-header';
+
+function ResizeIcon() {
+  return (
+    <svg
+      xmlns="http://www.w3.org/2000/svg"
+      width="20"
+      height="20"
+      viewBox="0 0 24 24"
+      strokeWidth="2"
+      stroke="#ff0071"
+      fill="none"
+      strokeLinecap="round"
+      strokeLinejoin="round"
+      style={{ position: 'absolute', right: 5, bottom: 5 }}
+    >
+      <path stroke="none" d="M0 0h24v24H0z" fill="none" />
+      <polyline points="16 20 20 20 20 16" />
+      <line x1="14" y1="14" x2="20" y2="20" />
+      <polyline points="8 4 4 4 4 8" />
+      <line x1="4" y1="4" x2="10" y2="10" />
+    </svg>
+  );
+}
+
+const controlStyle = {
+  background: 'transparent',
+  border: 'none',
+};
 
 export function IterationNode({
   id,
@@ -20,11 +48,14 @@ export function IterationNode({
   );
   return (
     <section
-      className={cn('bg-white w-full h-full', {
+      className={cn('w-full h-full', {
         [styles.dark]: theme === 'dark',
         [styles.selectedNode]: selected,
       })}
     >
+      <NodeResizeControl style={controlStyle} minWidth={100} minHeight={50}>
+        <ResizeIcon />
+      </NodeResizeControl>
       <Handle
         id="c"
         type="source"
@@ -79,7 +110,6 @@ export function IterationStartNode({
         position={Position.Right}
         isConnectable={isConnectable}
         className={styles.handle}
-        id="b"
         style={RightHandleStyle}
       ></Handle>
       <div>
