+++ web/src/pages/flow/form/invoke-form/dynamic-variables.tsx
@@ -2,11 +2,11 @@ import { EditableCell, EditableRow } from '@/components/editable-cell';
 import { useTranslate } from '@/hooks/common-hooks';
 import { DeleteOutlined } from '@ant-design/icons';
 import { Button, Collapse, Flex, Input, Select, Table, TableProps } from 'antd';
-import { useBuildComponentIdSelectOptions } from '../../hooks';
+import { trim } from 'lodash';
+import { useBuildComponentIdSelectOptions } from '../../hooks/use-get-begin-query';
 import { IInvokeVariable } from '../../interface';
 import { useHandleOperateParameters } from './hooks';
 
-import { trim } from 'lodash';
 import styles from './index.less';
 
 interface IProps {
