+++ web/src/pages/flow/index.tsx
@@ -5,7 +5,8 @@ import { ReactFlowProvider } from 'reactflow';
 import FlowCanvas from './canvas';
 import Sider from './flow-sider';
 import FlowHeader from './header';
-import { useCopyPaste, useFetchDataOnMount } from './hooks';
+import { useCopyPaste } from './hooks';
+import { useFetchDataOnMount } from './hooks/use-fetch-data';
 
 const { Content } = Layout;
